requirements.txt
django==2.2
djangorestframewor==3.12.2
requests==2.25.1
PyJWT==1.7.1
tenacity==5.1.1
djangorestframework==3.12.2